+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix crash on mapping/injection screens (`Safe\unserialize` not shipped by GLPI core yet)
+
 ## [2.15.9] - 2026-08-04
 
 ### Fixed